Project-2
Paris Olympic Games website

This page shows the medal ranking, sports and schedules , and the top news.. To get to every option on the menu you must click on the menu button that is on the top left corner and click on your option.

Medal Ranking : you will see the ranking of gold ,sliver, and bronze  of the top 10 countries

Sport/Schedule: you will see the top 5 sports and the days and times they are going to be showing

News: you will see the top news in media and on the internet that have been happen.